Relativity is a conspiracy
I changed it to deal with residuals instead of absolute coordinates. The fit functions for longitude and altitude had to change. The fit still doesn't work super well. I suspect that the assumption that the phase shift is the only thing that changes from variation to variation (ignoring the amplitude, offset, and frequency) is bogus.