Okay, but just one
I realized that we actually need the clock to be accurate for the residual computation thing to work at all. So we'll need to be correcting it as we go, anyway. Oh, well.

That's why I reverted the last commit. This one is to fix a bug I fixed in the commit I just reverted.